Fix (permute): refactor permutations into new permute.py
