+++ src/brevitas/graph/equalize.py
@@ -25,7 +25,6 @@ import torch
 from torch.fx import GraphModule as TorchGraphModule
 import torch.nn as nn
 import torch.nn.utils.parametrize as parametrize
-from tqdm import tqdm
 
 from brevitas import torch_version
 from brevitas.fx import GraphModule
@@ -43,7 +42,6 @@ from brevitas.graph.hadamard import get_hadK
 from brevitas.graph.hadamard import matmul_hadU
 from brevitas.graph.hadamard import matmul_hadU_cuda
 from brevitas.graph.hadamard import random_hadamard_matrix
-from brevitas.graph.utils import find_node_for_module
 from brevitas.graph.utils import get_module
 from brevitas.graph.utils import get_node
 from brevitas.nn import ScaledDotProductAttention
@@ -400,218 +398,6 @@ def apply_rewriters_accelerate(
     return model
 
 
-class rotate_permute_mode:
-
-    def __init__(self, model, permute_fn='massdiff', block_rotation_dim=None, **kwargs):
-        # rotate_permute_mode performs a specific transformation sequence:
-        # 1. Identify regions → 2. Apply permutations → 3. Apply rotations
-        # This sequencing requires delay_rewriters=True (defer rotation application)
-        # and return_rewriters=True (to access rotations for later application)
-        if not kwargs.pop('delay_rewriters', True) or not kwargs.pop('return_rewriters', True):
-            warnings.warn(
-                "delay_rewriters and return_rewriters must be True for rotate_permute_mode, ",
-                "overwriting provided value(s).")
-        kwargs['delay_rewriters'] = True
-        kwargs['return_rewriters'] = True
-
-        # NOTE: permutations are tied to block rotations here
-        self.rotation = GraphRotationEqualization(block_rotation_dim=block_rotation_dim, **kwargs)
-        self.permutation = GraphPermutationEqualization(block_rotation_dim=block_rotation_dim)
-        self.model = model
-        self.rewriters = None
-        self.permute_fn = get_permutation_method(permute_fn)
-
-    def __enter__(self):
-        model, rewriters = self.rotation.apply(self.model)
-        self.model = model
-        self.rewriters = rewriters
-
-        # Filter regions for permutation
-        permute_regions = list()
-        for region in self.rotation.regions:
-            # Permutations are only applied to regions that use block rotations
-            apply_block_rotation = self.permutation.block_rotation_dim is not None
-            if self.rotation.disable_block_rotation_for_fused and (len(region.srcs) > 0):
-                apply_block_rotation = False
-            if apply_block_rotation:
-                # Check if block rotation is compatible with the current shape
-                if (region.max_shape_sinks // self.permutation.block_rotation_dim > 1) and \
-                    (region.max_shape_sinks % self.permutation.block_rotation_dim == 0):
-                    permute_regions.append(region)
-
-        self.permutation.extract_permute_regions(model, permute_regions)
-        self.permutation.setup_permute()
-        return self
-
-    def __exit__(self, *args, **kwargs):
-        self.permutation.apply_permute(self.permute_fn)
-        self.permutation.remove_hooks()
-
-
-class GraphPermutationEqualization:
-    """
-    A class for managing and applying permutations to a computational graph.
-
-    This class is designed to analyze and modify computational graphs by identifying
-    regions of interest, collecting statistics, and applying permutations to optimize
-    or modify the graph's behavior. It supports various neural network layers and
-    operations, and provides hooks for collecting forward pass statistics.
-    """
-
-    def __init__(self, block_rotation_dim: int):
-        super().__init__()
-        self.hooks = []
-        self.hooked_modules = set()
-        self.supported_srcs = (nn.Embedding, RotatedModule, nn.Linear)
-        self.supported_sinks = (nn.Linear, RotatedModule)
-        common_scale_invariant = list(_scale_invariant_layers)
-        common_scale_invariant.extend([torch.nn.GELU, torch.nn.SELU])
-        scale_invariant_functions = (torch.nn.functional.silu,)
-        mul_ops = [torch.mul, operator.mul, operator.imul, operator.__mul__, operator.__imul__]
-        self.residual_fns = list(_residual_fns)
-        self.residual_fns.extend(mul_ops)
-        try:
-            # Add HuggingFace activations if available
-            from transformers.activations import ACT2CLS
-            activations = [x if not isinstance(x, tuple) else x[0] for x in ACT2CLS.values()]
-            common_scale_invariant.extend(activations)
-        except:
-            pass
-        self.scale_invariant_layers = tuple(common_scale_invariant) + (RMSNorm,)
-        self.scale_invariant_functions = tuple(scale_invariant_functions)
-        self.regions = list()
-        self.float_act_map = dict()
-        self.float_act_dev = dict()
-        self.block_rotation_dim = block_rotation_dim
-
-    def forward_stats_hook(self, module, *args, name, batch_dim=0, **kwargs):
-        # Check for MHA Cross attention, and if found, skip it
-        # When using hf/accelerate, we need to check the signature of the original forward
-        forward_to_check = module._old_forward if hasattr(
-            module, '_old_forward') else module.forward
-        kwargs.update(zip(forward_to_check.__code__.co_varnames[1:], args[:-1]))
-        if 'query' in kwargs and 'key' in kwargs and 'value' in kwargs:
-            if kwargs['query'].data_ptr() != kwargs['key'].data_ptr() != kwargs['value'].data_ptr():
-                return
-
-        inp_kwarg = [x for x in kwargs.keys() if x in INPUT_NAMES][0]
-        inp = kwargs[inp_kwarg][0]
-
-        # Extra check for batch_dim
-        if hasattr(inp, 'names') and 'N' in inp.names:
-            batch_dim = inp.names.index('N')
-            inp.rename_(None)
-            inp = inp.transpose(0, batch_dim)
-
-        inp = inp.reshape(-1, inp.shape[-1])  # [batch_size * seq_len, dim]
-        if name not in self.float_act_map:
-            self.float_act_map[name] = []
-            self.float_act_dev[name] = inp.device
-        self.float_act_map[name].append(inp.detach().cpu())
-
-    def setup_permute(self):
-        for region in self.regions:
-            # We assume that the entire region has a unique batch_dim
-            batch_dim = 0
-            for name in region.srcs:
-                module = region.get_module_from_name(name)
-                if hasattr(module, 'batch_first') and not module.batch_first:
-                    batch_dim = 1
-            for name in region.sinks:
-                module = region.get_module_from_name(name)
-                if hasattr(module, 'batch_first') and not module.batch_first:
-                    batch_dim = 1
-
-            for name in region.sinks_names:
-                module = region.get_module_from_name(name)
-                if module not in self.hooked_modules:
-                    self.hooked_modules.add(module)
-                    hook_fn = partial(self.forward_stats_hook, name=name, batch_dim=batch_dim)
-                    h = module.register_forward_hook(hook_fn)
-                    self.hooks.append(h)
-
-    def extract_permute_regions(self, graph_model, regions):
-        state_impl_kwargs = {
-            'supported_srcs': self.supported_srcs,
-            'supported_sinks': self.supported_sinks,
-            'scale_invariant_layers': self.scale_invariant_layers,
-            'scale_invariant_functions': self.scale_invariant_functions,
-            'residual_fns': self.residual_fns}
-
-        for region in regions:
-            # Directly add regions that already have sources identified
-            if (len(region.srcs) > 0):
-                # Skip the SDPA regions; potential head alignment issues
-                if 'value_sdpa' not in region.srcs_names:
-                    self.regions.append(region)
-                continue
-
-            # Create a new state for the online region
-            state = WalkRegionState(**state_impl_kwargs)
-
-            # Add all sinks from the region to the state
-            for sink_name, sink_wrapper in region.sinks.items():
-                module = region.get_module_from_name(sink_name)
-                node = find_node_for_module(graph_model, module)
-                assert node is not None, f"Error: node {module} not found in graph"
-                eq_indexes = sink_wrapper.equalization_indexes
-                state.add_sinks(node.target, module, eq_indexes)
-                find_srcs(graph_model, node, state)
-
-            # Create a new region with updated sources but same sinks
-            new_region = Region.from_dicts(
-                srcs=state.srcs,
-                sinks=state.sinks,
-                name_to_module=state.name_to_module,
-                expand_region=region.expand_region)
-            self.regions.append(new_region)
-
-    @staticmethod
-    def permute_region(region, list_of_act_val, block_rotation_dim, permute_fn, device):
-        """
-        Apply permutation to a region by calculating permutation indexes and updating
-        the source and sink weights accordingly.
-        """
-        # If equalization criteria are not met, return without doing anything
-        if not region.is_valid_activation_equalization:
-            return
-
-        list_of_act_val_shapes = [act_val.shape for act_val in list_of_act_val]
-        if len(list_of_act_val_shapes) > 0:
-            shape_0 = list_of_act_val_shapes[0]
-            if any(shape_0 != shape for shape in list_of_act_val_shapes):
-                return
-
-        list_of_act_val = torch.cat(list_of_act_val, dim=0).to(device)
-        new_indexes = permute_fn(list_of_act_val, block_rotation_dim=block_rotation_dim)
-
-        for src in region.srcs.values():
-            src.permute(new_indexes)
-        for sink in region.sinks.values():
-            sink.permute(new_indexes)
-
-    def apply_permute(self, permute_fn=None):
-        for region in tqdm(self.regions, "Calculating permutations..."):
-            # Collect all activation values for this region
-            list_of_act_val = []
-            for name in region.sinks_names:
-                act_vals = self.float_act_map.pop(name)
-                if act_vals is None or len(act_vals) == 0:
-                    continue
-                list_of_act_val.extend(act_vals)
-            # Calculate permutation and apply to this region
-            self.permute_region(
-                region,
-                list_of_act_val=list_of_act_val,
-                block_rotation_dim=self.block_rotation_dim,
-                permute_fn=permute_fn,
-                device=self.float_act_dev[region.sinks_names[0]])
-
-    def remove_hooks(self):
-        for h in self.hooks:
-            h.remove()
-
-
 class activation_equalization_mode:
 
     def __init__(
@@ -675,103 +461,6 @@ def _channel_range(inp: torch.Tensor, dim: int = 1) -> torch.Tensor:
 def _channel_maxabs(inp: torch.Tensor, dim: int = 1) -> torch.Tensor:
     out = torch.max(torch.abs(inp), dim=dim)[0]
     return out
-
-
-# Dictionary to store registered permutation methods
-_PERMUTATION_METHODS = {}
-
-
-def register_permutation_method(name: str):
-    """Register a permutation method for block rotations.
-
-    Args:
-        name: The name of the permutation method (e.g., "zigzag", "massdiff")
-
-    Examples:
-        >>> @register_permutation_method("my_permute")
-        ... def my_permute_method(x, block_rotation_dim):
-        ...     return torch.arange(x.shape[-1])
-    """
-
-    def _wrapper(permute_fn):
-        if name in _PERMUTATION_METHODS:
-            logging.warning(
-                "The permutation method '%s' already exists and will be "
-                "overwritten by %s.",
-                name,
-                permute_fn.__name__,
-            )
-        _PERMUTATION_METHODS[name] = permute_fn
-        return permute_fn
-
-    return _wrapper
-
-
-def get_permutation_method(name: str):
-    """Get a registered permutation method by name."""
-    if name not in _PERMUTATION_METHODS:
-        available = list(_PERMUTATION_METHODS.keys())
-        raise ValueError(
-            f"Permutation method '{name}' not found. "
-            f"Available methods: {available}")
-    return _PERMUTATION_METHODS[name]
-
-
-@register_permutation_method("zigzag")
-def zigzag_permute(x, block_rotation_dim):
-    if x.shape[-1] == block_rotation_dim:
-        return torch.arange(block_rotation_dim).to(x.device)
-    scores = _channel_maxabs(x, dim=0)
-    _, indexes = torch.sort(scores, descending=True)
-    # Inline zigzag sort logic
-    indexes = indexes.view(block_rotation_dim, indexes.shape[-1] // block_rotation_dim)
-    indexes[1::2] = torch.flip(indexes[1::2], dims=[1])
-    indexes = indexes.t()
-    indexes = indexes.flatten()
-    return indexes
-
-
-@register_permutation_method("random")
-def random_permute(x, block_rotation_dim):
-    if x.shape[-1] == block_rotation_dim:
-        return torch.arange(block_rotation_dim).to(x.device)
-    indexes = torch.randperm(x.shape[-1]).to(x.device)
-    return indexes
-
-
-@register_permutation_method("absmax")
-def absmax_permute(x, block_rotation_dim):
-    if x.shape[-1] == block_rotation_dim:
-        return torch.arange(block_rotation_dim).to(x.device)
-    scores = _channel_maxabs(x, dim=0)
-    _, indexes = torch.sort(scores, descending=True)
-    return indexes
-
-
-@register_permutation_method("massdiff")
-def massdiff_permute(x, block_rotation_dim):
-    if x.shape[-1] == block_rotation_dim:
-        return torch.arange(block_rotation_dim).to(x.device)
-    # initialize the blocks based on absmax scores
-    scores = torch.abs(x).mean(dim=0)
-    _, indexes = torch.sort(scores, descending=True)
-    num_blocks = x.shape[-1] // block_rotation_dim
-    # initialize the block norms and indexes
-    block_norm = torch.stack([torch.abs(x[:, i]) for i in indexes[:num_blocks]], dim=1)
-    block_idxs = [[i] for i in indexes[:num_blocks]]
-    for i in indexes[num_blocks:]:
-        # find the block that will have the minimum l1-norm after adding the new index
-        norms_after_adding = block_norm + torch.abs(x[:, i]).unsqueeze(1)
-        norms_after_adding = torch.mean(norms_after_adding, dim=0)
-        min_block = torch.argmin(norms_after_adding)
-        # update the block norm and indexes
-        block_norm[:, min_block] += torch.abs(x[:, i])
-        block_idxs[min_block].append(i)
-        # mark block as full
-        if (len(block_idxs[min_block]) == block_rotation_dim):
-            block_norm[:, min_block] = float('inf')
-    indexes = torch.tensor(block_idxs).flatten()
-    return indexes
 
 
 def _get_input_axis(module: nn.Module) -> Optional[int]:
